Should resolve the issues related to the <a> links created by pull request : cliftonc/calipso/pull#261. Changes: (1) - Dirty hack inside the link template rendering method that URI encodes parameters (2) - Replacing every mention of a verbose field (username, author, ...etc.) by it's URI-encoded copy (3) - Replacing every concatenation of variables to create URIs for HTTP redirects (res.redirect()) by their URI-encoded version
Together (1), (2) and (3) solve the issue raised in pull request #261 where <a> links were broken.
A quicker and cleaner way to fix this issue would be to create two hooks one as a middleware / prototype of Express.js's res.redirect() where the URI is encoded. The other hook would be called when creating <a> links when rendering templates so that every href is URI encoded.